+++ Practica3/product/src/main/java/com/product/api/controller/CtrlCategory.java
@@ -1,5 +1,5 @@
 package com.product.api.controller;
-
+import com.product.api.entity.Category;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.web.bind.annotation.RestController;
@@ -8,16 +8,15 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.ResponseBody;
 import java.util.LinkedList;
 import java.util.HashMap;
-import com.product.api.entity.CategoryManager.Category;
-
+import com.product.api.entity.CategoryManager;
 /**
 *Clase RestController para el control del envío respuesta
 * a request relacionados con categoría
 */
 @RestController
 public class CtrlCategory{
-	private  LinkedList<HashMap<String, Object>> api_list=new LinkedList<>();
-	private  LinkedList<Category> categories_list= new LinkedList<>();
+	private  static LinkedList<HashMap<String, Object>> api_list=new LinkedList<>();
+	private  static LinkedList<Category> categories_list= new LinkedList<>();
 
 
 	/**
@@ -30,7 +29,7 @@ public class CtrlCategory{
 	* ]
 	*
 	*/
-	CtrlCategory(LinkedList<Category> categories){
+	public CtrlCategory(LinkedList<Category> categories){
 		this.categories_list=categories;
 		fromCatToAPI();
 	}
@@ -43,6 +42,7 @@ public class CtrlCategory{
 	public  void fromCatToAPI(){
 
 		for(Category c: categories_list){
+			//System.out.println(c.toString());
 			api_list.add(new HashMap<String, Object>() {
 				{
 					put("category_id", c.getCategory_id());
@@ -51,22 +51,28 @@ public class CtrlCategory{
 			});
 		}
 	}
+
 	/**
 	*Devuelve una lista con el contenido de la API
 	*
 	*/
 	@GetMapping("/category")
 	public LinkedList<HashMap<String, Object>> getCategory(){
-		this.fromCatToAPI();
+		System.out.println("==categories list==");
+		for(Category c: categories_list){
+			System.out.println(c.toString());
+		}
 		return api_list;
 		//return new String[]{"categoría", "pollos", "verduras"};
 	}
 	@GetMapping("/category/{id}")
 	@ResponseBody
 	public LinkedList<HashMap<String, Object>> getCategoryById(@PathVariable Integer id){
+		boolean flag=false;
 		LinkedList<HashMap<String, Object>> local= new LinkedList<>();
 		for(Category c: categories_list){
 			if(c.getCategory_id().equals(id)){
+				flag=!flag;
 				local.add(new HashMap<String, Object>(){
 					{
 						put("category_id", c.getCategory_id());
@@ -75,8 +81,10 @@ public class CtrlCategory{
 					}
 				});
 			}
+			if(flag){
+				return local;
+			}
 
-					return local;
 		}
 
 		local.add(new HashMap<String, Object>(){
